requirements.txt
PyMuPDF
tika
python-dateutil
openpyxl
six
pandas
python-docx
